requirements.txt
fastapi
uvicorn
python-dotenv
bcrypt